move LAST_ISSUE into shared constants for HLR-996 and NOD-10182
