+++ src/applications/appeals/shared/utils/focus.js
@@ -5,7 +5,7 @@ import {
   waitForRenderThenFocus,
 } from 'platform/utilities/ui';
 import { $ } from 'platform/forms-system/src/js/utilities/ui';
-import { LAST_ISSUE } from '../../996/constants';
+import { LAST_ISSUE } from '../constants';
 
 export const focusIssue = (_index, root, value) => {
   setTimeout(() => {
